+++ src/commands/connections.ts
@@ -3,23 +3,29 @@ import { registerCommandWithLogging } from ".";
 import { getCCloudAuthSession } from "../authn/utils";
 import { openDirectConnectionForm } from "../directConnect";
 import { DirectConnectionManager } from "../directConnectManager";
+import { ccloudAuthSessionInvalidated } from "../emitters";
 import { Logger } from "../logging";
 import { DirectEnvironment } from "../models/environment";
 import { SSL_PEM_PATHS } from "../preferences/constants";
+import { deleteCCloudConnection } from "../sidecar/connections";
 import { CustomConnectionSpec, getResourceManager } from "../storage/resourceManager";
 import { ResourceViewProvider } from "../viewProviders/resources";
 
 const logger = new Logger("commands.connections");
 
 /** Allow CCloud sign-in via the auth provider outside of the Accounts section of the VS Code UI. */
-async function createCCloudConnection() {
+async function ccloudSignIn() {
   try {
     await getCCloudAuthSession(true);
   } catch (error) {
-    logger.error("error creating CCloud connection", { error });
     if (error instanceof Error) {
-      // if the user clicks "Cancel" on the modal before the sign-in process, we don't need to do anything
-      if (error.message === "User did not consent to login.") {
+      // we don't need to do anything if:
+      // - the user clicks "Cancel" on the modal before the sign-in process
+      // - the auth provider handles a callback failure (which shows its own error notification)
+      if (
+        error.message === "User did not consent to login." ||
+        error.message.includes("Authentication failed, see browser for details")
+      ) {
         return;
       }
       // any other errors will be caught by the error handler in src/commands/index.ts as part of the
@@ -27,6 +33,35 @@ async function createCCloudConnection() {
       throw error;
     }
   }
+}
+
+async function ccloudSignOut() {
+  const authSession = await getCCloudAuthSession();
+  if (!authSession) {
+    return;
+  }
+
+  // the authentication API doesn't provide a way to sign out, so we'll mirror the confirmation
+  // dialog from the Accounts section of the VS Code UI
+  const yesButton = "Sign Out";
+  const confirmation = await window.showInformationMessage(
+    `The account '${authSession.account.label}' has been used by: 
+
+Confluent
+
+Sign out from this extension?`,
+    {
+      modal: true,
+    },
+    yesButton,
+    // "Cancel" is added by default
+  );
+  if (confirmation !== yesButton) {
+    return;
+  }
+  // sign out by clearing the stored auth session
+  await deleteCCloudConnection();
+  ccloudAuthSessionInvalidated.fire();
 }
 
 /** Show the Open File dialog to let the user pick a .pem file and store it in the extension configs. */
@@ -143,7 +178,8 @@ export async function editDirectConnection(item: DirectEnvironment) {
 
 export function registerConnectionCommands(): Disposable[] {
   return [
-    registerCommandWithLogging("confluent.connections.ccloud.logIn", createCCloudConnection),
+    registerCommandWithLogging("confluent.connections.ccloud.signIn", ccloudSignIn),
+    registerCommandWithLogging("confluent.connections.ccloud.signOut", ccloudSignOut),
     registerCommandWithLogging("confluent.connections.addSSLPemPath", addSSLPemPath),
     registerCommandWithLogging("confluent.connections.direct", createNewDirectConnection),
     registerCommandWithLogging("confluent.connections.direct.delete", deleteDirectConnection),