+++ src/extension.ts
@@ -283,7 +283,6 @@ async function setupContextValues() {
     "ccloud-schema-registry",
     "local-kafka-cluster",
     "local-schema-registry",
-    "direct-kafka-cluster",
     "direct-schema-registry",
   ]);
   await Promise.all([