+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this extension will be documented in this file.
 
 - Producing messages to a Kafka topic has been expanded to allow loading content from an unsaved
   editor or Message Viewer preview tab.
+- "Sign Out" action in the Resources view for the current Confluent Cloud connection.
 
 ### Changed